Gateway/requirements.txt
fastapi
uvicorn
python-multipart
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
pytest
pytest-asyncio
httpx
websockets
Pillow
requests
PyYAML
jsonschema
onnxruntime
onnx
flask
